add schema for agent configuration files. Add tests. Refactor cli/common module.
