+++ aea/cli/helpers/schema.py
@@ -1,0 +1,38 @@
+# -*- coding: utf-8 -*-
+# ------------------------------------------------------------------------------
+#
+#   Copyright 2018-2019 Fetch.AI Limited
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+# ------------------------------------------------------------------------------
+
+"""Definition of the schemas for configuration file."""
+"""
+aea_version: 0.1.1
+agent_name: myagent
+connections:
+- connection:
+    config:
+      config:
+        addr: 127.0.0.1
+        port: 10000
+    name: default-oef
+    type: oef
+default_connection: default-oef
+protocols: []
+skills: []
+
+"""
+
+
