+++ aea/cli/remove.py
@@ -24,7 +24,7 @@ import shutil
 
 import click
 
-from aea.cli.common import DEFAULT_AEA_CONFIG_FILE, AgentConfig, Context, pass_ctx, logger
+from aea.cli.helpers.common import DEFAULT_AEA_CONFIG_FILE, AgentConfig, Context, pass_ctx, logger
 
 
 @click.group()
