+++ aea/cli/schemas/aea-config_schema.json
@@ -1,0 +1,42 @@
+{
+    "$schema": "http://json-schema.org/draft-07/schema#",
+    "description": "Schema for the agent configuration file.",
+    "additionalProperties": false,
+    "type": "object",
+    "required": ["aea_version", "agent_name", "connections", "default_connection", "protocols", "skills"],
+    "properties": {
+        "aea_version": {"type": "string"},
+        "agent_name": {"type": "string"},
+        "connections": {
+            "type": "array",
+            "uniqueItems": true,
+            "items": {
+              "type": "object",
+              "required": ["connection"],
+              "properties": {
+                "connection": { "$ref": "#/definitions/connection" }
+              }
+            }
+        },
+        "default_connection": {"type": "string"},
+        "protocols": {
+            "type": "array",
+            "items": {"type": "string"}
+        },
+        "skills": {
+            "type": "array",
+            "items": {"type": "string"}
+        }
+    },
+    "definitions": {
+        "connection": {
+            "type": "object",
+            "properties": {
+                "name": {"type": "string"},
+                "type": {"type": "string"},
+                "config": {"type": "object"}
+            },
+            "required": ["name", "type", "config"]
+        }
+    }
+}
