+++ tests/test_cli.py
@@ -18,10 +18,16 @@
 # ------------------------------------------------------------------------------
 
 """This test module contains the tests for the `aea` command-line tool."""
+import json
+import os
+import pprint
 
-
+import yaml
 from click.testing import CliRunner
+from jsonschema import validate, Draft7Validator
+
 from aea.cli import cli
+from .conftest import CUR_PATH, ROOT_DIR
 
 
 def test_no_argument():
@@ -29,3 +35,17 @@ def test_no_argument():
     runner = CliRunner()
     result = runner.invoke(cli, [])
     assert result.exit_code == 0
+
+
+def test_agent_configuration_schema_is_valid_wrt_draft_07():
+    """Test that the JSON schema for the agent configuration file is compliant with the specification Draft 07."""
+    agent_config_schema = json.load(open(os.path.join(ROOT_DIR, "aea", "cli", "schemas", "aea-config_schema.json")))
+    Draft7Validator.check_schema(agent_config_schema)
+
+
+def test_validate_config():
+    """Test that the validation of the agent configuration file works correctly."""
+    agent_config_schema = json.load(open(os.path.join(ROOT_DIR, "aea", "cli", "schemas", "aea-config_schema.json")))
+    agent_config_file = yaml.safe_load(open(os.path.join(CUR_PATH, "data", "aea-config.example.yaml")))
+    pprint.pprint(agent_config_file)
+    validate(instance=agent_config_file, schema=agent_config_schema)
